requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.5.0
astroid==2.11.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
dj-rest-auth==2.2.7
Django==3.2.7
django-allauth==0.52.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
filelock==3.6.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
msgpack==1.0.3
msgpack-python==0.5.6
mypy-extensions==0.4.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.22.4
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.4.0
platformdirs==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.13.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==2.10.6
requests==2.27.1
requests-oauthlib==1.3.1
scipy==1.8.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.0
wrapt==1.14.0
zope.interface==5.4.0
